requirements.txt
pyrogram
tgcrypto
python-telegram-bot
